implement new mechanism to resolve a object name
Sometimes the exact object name of a Mbean cannot be determined beforehand.
This commit implements a mechanism to resolve "object name pattern"

The Kafka request counter required a "version" key in the object
name query. Which stands for the version of a specific RPC request
format.

If we hard code the value in code when the version changed we need to
modify it manually. Which cause maintenance problem.

There is actually an API on the MBean server that can be used to lookup
object name in a programmatic way. We used this API with a
regular-expression-like object name to search for the exact object name.

This commit takes advantage of that to resolve a object name.