fix incorrect metric type
The metric "kafka_server_kafkarequesthandlerpool_requesthandleravgidle_percent"
was set to type "counter". Which is incorrect since it is a value
reflect the current state instead of a value increment over times.

This commit hack this issue by update the config for jmx exporter. The
new config will set that metric as "gauge" type. Which is more suitable
for reflect the current state of a metric.