Chagne the way we initialize a `ClusterLogAllocation`
Originally we converted the content of a `ClusterInfo` into a map
structure that maps the replica list of each topic/partition. Now we
initialize from a list of `Replica`. Entirely get rid of
`LogPlacement`.

In theory, this will avoid the extensive conversion of POJO between the
iteration of ClusterLogAllocation with other classes.