avoid client missing metrics.
Before this commit,
client can read data like this closed interval [since, threadTime]

This commit change it to a half-open interval [since, threadTime)

It is possible that multiple threads attempt to put data into the
threadTime bucket. Which cause some unfortunate client to access
partical data at that point of millisecond.

This commit make it impossible to read data at that point of time.
Also perform a intentional sleep, which cause the updated thread time
become (threadTime + 1) instead of just (threadTime). This effectively
forbidden the processing thread to create any work at that moment. So
the fresh metrics it got can be safely published.