[EXPORTER] Add metadata in record writer (#1832)
* recordWriter will append metadata before closing.

During the creation of a new recordWriter, the exporter task will transmit both the configuration and the first record for writing metadata. A new method has been added to the SinkContext for retrieving the connector name to be used within the metadata.

* a new version of recordWriter v1, which contains metadata at the end of a file.

The reason for creating a new version instead of replacing version 0 is because the importer is not yet able to handle files with metadata information.

* Fix RecordReader to not error out when receiving V1 files

While it can accept V1 files, it was only processing the initial data and not the following metadata

* Remove topic/partition/1st record info metadata

Removing these data from metadata to give writers more flexibility. This allows writers to write data without restrictions, allowing the exporter to decide how they want to handle the data.